+++ src/vanilla/injector.test.ts
@@ -1,6 +1,11 @@
 import { atom, PrimitiveAtom } from "jotai";
 import { createInjector, getDefaultInjector } from "./injector";
-import { ErrorInvalidGlobalInjector, ErrorInvalidMolecule, ErrorInvalidScope, ErrorUnboundMolecule } from "./internal/errors";
+import {
+  ErrorInvalidGlobalInjector,
+  ErrorInvalidMolecule,
+  ErrorInvalidScope,
+  ErrorUnboundMolecule,
+} from "./internal/errors";
 import { DefaultInjector } from "./internal/symbols";
 import { Molecule, molecule, moleculeInterface } from "./molecule";
 import { createScope } from "./scope";
@@ -14,8 +19,6 @@ const exampleMol = molecule<BaseAtoms>(() => {
     nameAtom: atom(`${Math.random()}`),
   };
 });
-
-
 
 const UnrelatedScope = createScope<number>(1);
 const unrelatedScope1: ScopeTuple<number> = [UnrelatedScope, 1];
@@ -64,7 +67,6 @@ test("returns the same values for dependency-free molecule", () => {
 });
 
 describe("Derived molecules", () => {
-
   const derivedMol = molecule((mol) => {
     const base = mol(exampleMol);
     return { base, ageAtom: atom(`${Math.random()}`) };
@@ -104,8 +106,7 @@ describe("Derived molecules", () => {
   test("returns the same value for 2nd order derived molecule", () => {
     testDerived(doubleDerived);
   });
-})
-
+});
 
 test("two injectors return different molecules", () => {
   const injector1 = createInjector();
@@ -417,11 +418,7 @@ describe("Scoping", () => {
 
     test("Works with a diamond pattern dependency tree, with sibling dependency", () => {
       const molLeft = molecule((mol) => ["left", mol(molTop)]);
-      const molRight = molecule((mol) => [
-        "right",
-        mol(molTop),
-        mol(molLeft),
-      ]);
+      const molRight = molecule((mol) => ["right", mol(molTop), mol(molLeft)]);
 
       const molBottom = molecule((mol) => [
         "bottom",
@@ -538,48 +535,48 @@ describe("Scoping", () => {
 });
 
 describe("Validation", () => {
-
   test("Molecules will throw errors if `mol` is called asynchronously", async () => {
-
     const badMolecule = molecule((mol) => {
       // Okay -- runs sync
       mol(exampleMol);
-      return new Promise((resolve, reject) => setTimeout(() => {
-        try {
-          // Not okay -- runs in a timeout
-          resolve(mol(exampleMol))
-        } catch (e) {
-          reject(e);
-        }
-      }, 10));
+      return new Promise((resolve, reject) =>
+        setTimeout(() => {
+          try {
+            // Not okay -- runs in a timeout
+            resolve(mol(exampleMol));
+          } catch (e) {
+            reject(e);
+          }
+        }, 10)
+      );
     });
 
     const injector1 = createInjector();
 
     const firstValue = injector1.get(badMolecule);
     await expect(firstValue).rejects.toThrow("o");
-  })
-
+  });
 
   test("Molecules will throw errors if `scope` is called asynchronously", async () => {
-
     const badMolecule = molecule((_, scope) => {
       // Okay -- runs sync
-      return new Promise((resolve, reject) => setTimeout(() => {
-        try {
-          // Not okay -- runs in a timeout
-          resolve(scope(UserScope))
-        } catch (e) {
-          reject(e);
-        }
-      }, 10));
+      return new Promise((resolve, reject) =>
+        setTimeout(() => {
+          try {
+            // Not okay -- runs in a timeout
+            resolve(scope(UserScope));
+          } catch (e) {
+            reject(e);
+          }
+        }, 10)
+      );
     });
 
     const injector1 = createInjector();
 
     const firstValue = injector1.get(badMolecule);
     await expect(firstValue).rejects.toThrow("o");
-  })
+  });
 
   describe("Bad dependencies", () => {
     const injector1 = createInjector();
@@ -587,46 +584,46 @@ describe("Validation", () => {
     test("Molecules can't depend on garbage molecules", () => {
       const badMol = molecule((mol) => mol(new Set() as any));
       expect(() => injector1.get(badMol)).toThrow(ErrorInvalidMolecule);
-    })
+    });
     test("Molecules can't depend on garbage scopes", () => {
       const badMol = molecule((_, scope) => scope(new Set() as any));
       expect(() => injector1.get(badMol)).toThrow(ErrorInvalidScope);
-    })
+    });
   });
 
   describe("Validation for inputs to injector", () => {
     const injector1 = createInjector();
 
     test("Can't `get` a non-molecule", () => {
-      expect(() => injector1.get(new Set() as any)).toThrow(ErrorInvalidMolecule);
-    })
+      expect(() => injector1.get(new Set() as any)).toThrow(
+        ErrorInvalidMolecule
+      );
+    });
     test("Can't `use` a non-molecule", () => {
-      expect(() => injector1.use(new Map() as any)).toThrow(ErrorInvalidMolecule);
-    })
-
+      expect(() => injector1.use(new Map() as any)).toThrow(
+        ErrorInvalidMolecule
+      );
+    });
   });
-})
-
+});
 
 describe("Binding", () => {
-
   interface HTTPService {
     identity: string;
-    get(url: string): Promise<string>
-    post(url: string): Promise<string>
+    get(url: string): Promise<string>;
+    post(url: string): Promise<string>;
   }
 
   const HTTPService = moleculeInterface<HTTPService>();
 
   const NeedsHttp = molecule((mol) => {
-
     const httpService = mol(HTTPService);
 
-    const logout = () => httpService.post("/logout")
+    const logout = () => httpService.post("/logout");
     return {
       httpService,
-      logout
-    }
+      logout,
+    };
   });
 
   test("Errors when a molecule interface is not bound", () => {
@@ -635,10 +632,7 @@ describe("Binding", () => {
   });
 
   describe("Allows binding a molecule interface to a molecule", () => {
-
-
     const MockHTTPMolecule = molecule<HTTPService>(() => {
-
       return {
         identity: "MockHTTP",
         async get(url) {
@@ -647,12 +641,12 @@ describe("Binding", () => {
         async post(url) {
           return "I am fake";
         },
-      }
+      };
     });
 
     test("Injects bindings into downstream dependencies", () => {
       const injector1 = createInjector({
-        bindings: [[HTTPService, MockHTTPMolecule]]
+        bindings: [[HTTPService, MockHTTPMolecule]],
       });
 
       const firstValue = injector1.get(NeedsHttp);
@@ -660,49 +654,43 @@ describe("Binding", () => {
 
       const bound = injector1.get(HTTPService);
       expect(bound).toBe(firstValue.httpService);
-    })
+    });
 
     describe("Work with tuple bindings", () => {
       const arrayBindings = [[HTTPService, MockHTTPMolecule]];
       const injector1 = createInjector({
-        bindings: [[HTTPService, MockHTTPMolecule]]
+        bindings: [[HTTPService, MockHTTPMolecule]],
       });
 
       test("injects the right values", () => {
         expect(injector1.get(HTTPService).identity).toBe("MockHTTP");
-      })
+      });
       test("doesn't update the injector when the bindings array is mutated", () => {
         arrayBindings.pop();
-        expect(arrayBindings).toStrictEqual([])
+        expect(arrayBindings).toStrictEqual([]);
         expect(injector1.get(HTTPService).identity).toBe("MockHTTP");
-      })
-    })
+      });
+    });
 
     describe("Works with map bindings", () => {
-
       const mapBindings = new Map();
       mapBindings.set(HTTPService, MockHTTPMolecule);
       const injector1 = createInjector({
-        bindings: mapBindings
+        bindings: mapBindings,
       });
 
       test("injects the right values", () => {
         expect(injector1.get(HTTPService).identity).toBe("MockHTTP");
-      })
-
+      });
 
       test("doesn't update the injector when the map is mutated", () => {
         mapBindings.delete(HTTPService);
         expect(injector1.get(HTTPService).identity).toBe("MockHTTP");
-      })
-    })
-
-
-
-  })
+      });
+    });
+  });
 
   test("Allows binding a molecule interface to a scoped molecule", async () => {
-
     const UserScopedHTTPMolecule = molecule<HTTPService>((_, getScope) => {
       const user = getScope(UserScope);
       return {
@@ -713,11 +701,11 @@ describe("Binding", () => {
         async post(url) {
           return `I am ${user}`;
         },
-      }
+      };
     });
 
     const injector1 = createInjector({
-      bindings: [[HTTPService, UserScopedHTTPMolecule]]
+      bindings: [[HTTPService, UserScopedHTTPMolecule]],
     });
 
     const firstValue = injector1.get(NeedsHttp);
@@ -729,16 +717,20 @@ describe("Binding", () => {
 
     const loggedOut2 = await secondValue.logout();
     expect(loggedOut2).toBe("I am one@example.com");
-
-  })
-})
+  });
+});
 
 describe("Scope caching", () => {
-
   test("Caches a scope tuple", () => {
     const injector = createInjector();
-    const [[tuple1], unsub1] = injector.useScopes([UserScope, "one@example.com"]);
-    const [[tuple2], unsub2] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple1], unsub1] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
+    const [[tuple2], unsub2] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
     expect(tuple1).toBe(tuple2);
 
     unsub1();
@@ -746,63 +738,98 @@ describe("Scope caching", () => {
   });
 
   describe("Caches a scoped molecule", () => {
-
     const injector = createInjector();
 
     describe("String scopes", () => {
-      const UserScoped = molecule((_, scope) => scope(UserScope) + Math.random());
+      const UserScoped = molecule(
+        (_, scope) => scope(UserScope) + Math.random()
+      );
       test("It caches for overlapping leases", () => {
-        const [mol1, unsub1] = injector.use(UserScoped, [UserScope, "one@example.com"]);
-        const [mol2, unsub2] = injector.use(UserScoped, [UserScope, "one@example.com"]);
+        const [mol1, unsub1] = injector.use(UserScoped, [
+          UserScope,
+          "one@example.com",
+        ]);
+        const [mol2, unsub2] = injector.use(UserScoped, [
+          UserScope,
+          "one@example.com",
+        ]);
         expect(mol1).toBe(mol2);
 
         unsub1();
         unsub2();
-      })
+      });
 
       test("It does not cache when leases are not overlapping", () => {
-        const [mol1, unsub1] = injector.use(UserScoped, [UserScope, "one@example.com"]);
+        const [mol1, unsub1] = injector.use(UserScoped, [
+          UserScope,
+          "one@example.com",
+        ]);
         unsub1();
-        const [mol2, unsub2] = injector.use(UserScoped, [UserScope, "one@example.com"]);
+        const [mol2, unsub2] = injector.use(UserScoped, [
+          UserScope,
+          "one@example.com",
+        ]);
         unsub2();
 
         expect(mol1).not.toBe(mol2);
-      })
-    })
+      });
+    });
     describe("Objects scopes", () => {
       const objectScope = createScope(new Set());
 
-      const ObjectScopedMol = molecule((_, scope) => new Set(scope(objectScope).entries()));
+      const ObjectScopedMol = molecule(
+        (_, scope) => new Set(scope(objectScope).entries())
+      );
 
       test("It caches for overlapping leases", () => {
         const testSet = new Set();
-        const [mol1, unsub1] = injector.use(ObjectScopedMol, [objectScope, testSet]);
-        const [mol2, unsub2] = injector.use(ObjectScopedMol, [objectScope, testSet]);
+        const [mol1, unsub1] = injector.use(ObjectScopedMol, [
+          objectScope,
+          testSet,
+        ]);
+        const [mol2, unsub2] = injector.use(ObjectScopedMol, [
+          objectScope,
+          testSet,
+        ]);
         expect(mol1).toBe(mol2);
         unsub1();
         unsub2();
-      })
+      });
 
-      test("It cache EVEN WHEN leases are not overlapping", () => {
+      test("It does NOT cache when leases are not overlapping", () => {
+        // Note: Behaviour changed in Version 1.1
+
         const testSet = new Set();
-        const [mol1, unsub1] = injector.use(ObjectScopedMol, [objectScope, testSet]);
+        const [mol1, unsub1] = injector.use(ObjectScopedMol, [
+          objectScope,
+          testSet,
+        ]);
         unsub1();
-        const [mol2, unsub2] = injector.use(ObjectScopedMol, [objectScope, testSet]);
+        const [mol2, unsub2] = injector.use(ObjectScopedMol, [
+          objectScope,
+          testSet,
+        ]);
         unsub2();
 
         expect(mol1).toBe(mol2);
-      })
-    })
+      });
+    });
   });
 
   test("Does not cache when scopes are cleaned up", () => {
     const injector = createInjector();
-    const [[tuple1], unsub1] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple1], unsub1] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
     unsub1();
 
     // Note: GC / cleanup happens in here
 
-    const [[tuple2], unsub2] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple2], unsub2] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
     unsub2();
     // Subscription 1 and 2 never overlapped
     expect(tuple1).not.toBe(tuple2);
@@ -810,8 +837,14 @@ describe("Scope caching", () => {
 
   test("Caches if there are overlapping subscriptions", () => {
     const injector = createInjector();
-    const [[tuple1], unsub1] = injector.useScopes([UserScope, "one@example.com"]);
-    const [[tuple2], unsub2] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple1], unsub1] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
+    const [[tuple2], unsub2] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
     unsub2();
     unsub1();
     // Subscription 2 overlapped with 1
@@ -820,18 +853,29 @@ describe("Scope caching", () => {
 
   test("Caches as long as subscriptions overlap", () => {
     const injector = createInjector();
-    const [[tuple1], unsub1] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple1], unsub1] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
 
-
-    const [[tuple2], unsub2] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple2], unsub2] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
 
     // Doesn't create a new value, the second use has a lease
     unsub1();
 
-    const [[tuple3], unsub3] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple3], unsub3] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
     unsub2();
 
-    const [[tuple4], unsub4] = injector.useScopes([UserScope, "one@example.com"]);
+    const [[tuple4], unsub4] = injector.useScopes([
+      UserScope,
+      "one@example.com",
+    ]);
     unsub3();
 
     // Final cleanup
@@ -841,11 +885,9 @@ describe("Scope caching", () => {
     expect(tuple1).toBe(tuple3);
     expect(tuple1).toBe(tuple4);
   });
-})
-
+});
 
 describe("Global injector", () => {
-
   test("It always returns the same injector", () => {
     (globalThis as any)[DefaultInjector] = undefined;
 
@@ -858,7 +900,7 @@ describe("Global injector", () => {
 
     const three = getDefaultInjector();
     expect(three).not.toBe(two);
-  })
+  });
 
   test("It returns different objects if the global scope is nulled out", () => {
     (globalThis as any)[DefaultInjector] = undefined;
@@ -867,15 +909,17 @@ describe("Global injector", () => {
     (globalThis as any)[DefaultInjector] = undefined;
     const two = getDefaultInjector();
     expect(one).not.toBe(two);
-  })
+  });
 
   test("It throws errors when global scoped is poluted with garbage", () => {
     // Pollute global scope with garbage
-    (globalThis as any)[DefaultInjector] = { "I am a lot of": "Terrible garbage" };
+    (globalThis as any)[DefaultInjector] = {
+      "I am a lot of": "Terrible garbage",
+    };
 
     expect(getDefaultInjector).toThrow(ErrorInvalidGlobalInjector);
 
     // Cleanup garbage
     (globalThis as any)[DefaultInjector] = undefined;
-  })
-})
+  });
+});